requirements.txt
Django>=1.8.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
static3==0.5.1
wsgiref==0.1.2
-e git+https://github.com/funkybob/django-flatblocks.git#egg=django-flatblocks
djrill
pytest-django
pytest-xdist
psycopg2==2.5.4
django-user-accounts==1.0
django-uuidfield==0.5.0
django-queryset-csv==0.2.10
newrelic==2.36.0.30
pdfkit==0.4.1
django-hermes
mock
rq
celery>=3.1
django-celery
jsonfield
eventlet
django-debug-toolbar==1.2.1
ipdb==0.8
ipython==2.2.0
django_import_export
yolk==0.4.3
